[Upd] Mode home to teamname app
